Use passed caster-Hero for ActiveSpell-target
Previously, battle.hero1 was always used for the target, which made
sense except for when spells cast by the other player needed to be
applied. This adds a parameter so the ActiveSpell knows who cast it.

The reason this touches so many files is that the parameterization of
Hero<Spell<*>> needed to be removed because the type system isn't
capable enough. This means that using a Hero's Spell will always require
a check of the Spell type.

Closes #62