Apply spells immediately on cast
This fixes a crash when BattleController would pick a unit to target
using EphemeralAllegiance, and then that unit would move before the
spell was applied.

AdrenalineShot was correspondingly changed to apply on turnsSinceCast
= 1, and extended with 1 turn duration, to keep the past behavior.
EphemeralAllegiance was also extended with 1 duration because it's not
much help the first turn, since the caster presumably has used their
action points.

Closes #64